memcached library support
PHP supports two memcached libraries: memcache and (the poorly named)
memcached.  w3tc only supports the memcache library extension.  Because
the memcache library is still not supported under  php7 this only leaves
the memcached library as the only option to get memcached daemon
support.  This update checks for and makes use of the memcached
extension library if available, and thus, providing php7 support for a
memcached server.