PageSpeeds update & Premium/Licensing Widgets Removed
Included a feature to allow users to provide an optional http referrer
(via a new input field under General Settings) which can be used as a
key restriction parameter for the PageSpeed console's credentials area.

Fixed an error that was triggered by a pagespeed response; due to a
change in the pagespeed api, a field that was expected was no longer
available.

Removed the General Setting's Licensing widget and the Dashboard Premium
Services widget.