Fragment Caching
This resolves a minor issue when users implement <--mfunc but trigger a
ParseError exception when using php7.  Prior to php7 eval() returned
false when an error occurred.  w3tc would then display a friendly error
message on screen and resume.  As of php7 the code is halted on parse
errors.  This revision makes it so that when using php7 a parsing error
will act like previous php versions when using w3tc by informing the
user but continuing to process the page.