updater revision
Prior to this the update it would only alert the user of new versions
when logged in.  this update now works more in line with other wp
plugins when updates are available so that mail on update type plugins
(which email when updates are available) to be used to alert users when
not logged in too.