CDN Custom Files Issue with Full Urls
When full urls (e.g. http://mydomain.com/uploads/*) are used in the
custom file list box (of CDN tab) it causes the page to break.  It's
because the regex that handles full urls is not correctly implemented.
Relative path urls work correctly. On further inspection it appears the
CDN page is purposely designed for local files only. Knowing that, this
fix simply strips full urls down to their paths for processing.

This issue was discovered because the {uploads_dir} placeholder when
converted was becoming a full url (e.g. {uploads_dir} =
http://mydomain.com/uploads/*) instead of being just its path (e.g.
/uploads/*) like {wp_content_dir} and {plugins_dir}.  As such, this fix
resolves the {upload_dir}/* issue.