Flushing Not Working for Post Changes
It was discovered that while logged in and performing an action such as
deleting a post or changing it to draft status, that the post does not
flush despite the purge policy settings.

Why the Problem Occurs

There are two underlying causes:

1. Several WP actions such as delete_post and save_post (among many
others) appear to be missing.  These actions are necessary to trigger
W3TC to flush when a post changes.  These actions were previously used
in the v0.9.4.x generation but most appear to be left out (by accident?)
for v0.9.5.x.  These missing actions were added (to
Util_AttachToActions.php) and, in the process, fixed several problems.

2. In regards to "Draft" and "Pending Review" though, when a post is
switched from "Published" to "Draft" (or "Pending Review") W3TC is
unable to handle flushing because the returned URL from get_permalink(),
as used in W3TC's get_post_urls() function, is in plain form (i.e.,
http://foo.bar/?p=###) instead of being in the normal one that uses the
site's permalink setting (e.g., http://foo.bar/post-name/).  This seems
to only occur when dealing with draft or pending review status.  It also
appears to be an old bug that spans all versions of W3TC.  By setting
the "post_status" variable (of the $post object) to empty and passing
$post to get_permalink() instead of the post id, it returns the correct
URL (a URL that uses the site's permalink setting).  And with that,
flushing begins to work.