Rewrite URLs for Media Library
This is a revision to #336 that moves away from using the
wp_prepare_attachment_for_js() filter and uses wp_get_attachment_url and
attachment_link filters instead.

The change makes the code less intense since it no longer has to parse
through the large array response for urls to change.

This revision also includes a check to verify that, if logged in, the
role is permitted to use CDN.