Don't confuse different files with the same title in Google Drive CDN
The Google Drive CDN export code was assuming that files in the same
chunk with the same title were the same file. This was causing all
sorts of problems, such as attempting to delete files multiple times,
attempting to access files that had been deleted, uploading files that
had already been uploaded repeatedly, comparing local files with the
wrong uploaded files to determine whether they needed to be uploaded
again, etc.

I fixed this by adding a new "path" property to uploaded files with
the remote path in it, and using this new property to disambiguate
files with the same remote title.

Note that the first time an export is run after this patch is applied,
all files will be uploaded again, even files that haven't changed,
because they don't have the "path" property yet.